Fixed default style issue with SuperTextField
