+++ super_editor/lib/src/infrastructure/super_textfield/ios/ios_textfield.dart
@@ -1,7 +1,6 @@
 import 'package:attributed_text/attributed_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:super_editor/src/default_editor/super_editor.dart';
 import 'package:super_editor/src/infrastructure/_listenable_builder.dart';
 import 'package:super_editor/src/infrastructure/_logging.dart';
 import 'package:super_editor/src/infrastructure/attributed_text_styles.dart';
@@ -12,6 +11,7 @@ import 'package:super_editor/src/infrastructure/super_textfield/ios/_editing_con
 import 'package:super_text/super_text.dart';
 
 import '../../platforms/ios/toolbar.dart';
+import '../styles.dart';
 import '_floating_cursor.dart';
 import '_user_interaction.dart';
 
@@ -28,7 +28,7 @@ class SuperIOSTextField extends StatefulWidget {
     Key? key,
     this.focusNode,
     this.textController,
-    this.textStyleBuilder = defaultStyleBuilder,
+    this.textStyleBuilder = defaultTextFieldStyleBuilder,
     this.textAlign = TextAlign.left,
     this.hintBehavior = HintBehavior.displayHintUntilFocus,
     this.hintBuilder,
